backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
pydantic>=2.0,<3.0
google-api-python-client>=2.175.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
